Split subscription tests
